requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
DateTime==5.4
dj-database-url==2.1.0
Django==4.2.9
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-hosts==6.0
django-log-request-id==2.1.0
django-otp==1.3.0
django-softdelete==0.10.5
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docutils==0.20.1
gunicorn==20.1.0
idna==3.6
JSON-log-formatter==0.5.2
kombu==5.3.4
logger==1.4
nsepythonserver==2.7
numpy==1.26.2
pandas==2.1.4
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
scipy==1.11.4
six==1.16.0
sqlparse==0.4.4
statistics==1.0.3.5
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
whitenoise==6.6.0
zope.interface==6.1
